feat: Refactor Docker containers and volumes setup
The code changes involve restructuring the Docker containers and volumes setup. This includes creating a network, volumes for the DB and Mail containers, and updating the App container. The purpose of these changes is to improve the organization and management of the containers.

Recent user commits and repository commits indicate a consistent use of the "feat" prefix for feature additions and updates. Following this convention, the suggested commit message is "feat: Refactor Docker containers and volumes setup".